+++ x/config.go
@@ -26,7 +26,7 @@ type Options struct {
 	PortOffset     int
 	QueryEdgeLimit uint64
 	// NormalizeNodeLimit is the maximum number of nodes allowed in a normalize query.
-	NormalizeNodeLimit uint64
+	NormalizeNodeLimit int
 }
 
 var Config Options
